Added guns and targets
Imported a new asset called Weapon Pack (LOW POLY). Attached an AK-47 to the main camera. Added two scripts, PlayerShootingScript, which is attached to the main camera, and ShotAtScript, which is attached to three targets I added to the prison scene. I also changed the main camera a little bit.